% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_table}
\alias{get_table}
\title{Convert PDF Tables to format more amenable to analysis}
\usage{
get_table(file, format, api_key = Sys.getenv("pdftable_api"))
}
\arguments{
\item{file}{The PDF file to be converted}

\item{format}{One of 'csv', 'xlm', 'xlsx-single', 'xlsx-multiple'}

\item{api_key}{Your API key (from https://pdftables.com)}
}
\value{
A string (format = 'csv', or format = 'xml') or raw vector (format =
  'xlsx-single' or 'xlsx-multiple`)
}
\description{
Convert PDF Tables to format more amenable to analysis
}
\examples{
\dontrun{
write.csv(iris, file = "test.csv", row.names = FALSE)

# Open test.csv and print as PDF

test_xlsx <- get_table(file = "test.pdf", format = "xlsx-multiple")

f <- file("test.xlsx", "wb")
writeBin(test_xlsx, f)
close(f)
}
}

